zzzz

+++ src/lib.rs
@@ -499,18 +499,21 @@ pub fn input_field_scrollable(history: &mut InputHistory) -> String{
             print!("{}", item);
         }
     }
-    fn scroll_action(res: &mut Vec<char>, newstr: &str, pos: &mut usize){
-        let old_len = delete_all(res);
-        feed_into_buffer(res, newstr);
-        write_all(&res);
-        *pos = res.len();
-        let diff = old_len as i32 - res.len() as i32;
-        if diff <= 0 { return; }
-        for _ in 0..diff{
-            print!(" ");
-        }
-        for _ in 0..diff{
-            print!("{}", 8 as char);
+    fn scroll_action(res: &mut Vec<char>, pos: &mut usize, history: &InputHistory, his_index: i32){
+        let val = history.get_index(his_index);
+        if let Some(valv) = val{
+            let old_len = delete_all(res);
+            feed_into_buffer(res, valv);
+            write_all(&res);
+            *pos = res.len();
+            let diff = old_len as i32 - res.len() as i32;
+            if diff <= 0 { return; }
+            for _ in 0..diff{
+                print!(" ");
+            }
+            for _ in 0..diff{
+                print!("{}", 8 as char);
+            }
         }
     }
     fn delete(res: &mut Vec<char>, pos: &mut usize, gstate: &mut u8){
@@ -576,10 +579,7 @@ pub fn input_field_scrollable(history: &mut InputHistory) -> String{
                 if gstate == 2 {
                     gstate = 0; 
                     his_index += 1;
-                    let val = history.get_index(his_index);
-                    if let Some(valv) = val{
-                        scroll_action(&mut res, valv, &mut pos);
-                    }
+                    scroll_action(&mut res, &mut pos, &history, his_index);
                 }
                 else { typed_char(65, &mut res, &mut gstate, &mut hoen_state, &mut pos); }
             }
@@ -587,10 +587,7 @@ pub fn input_field_scrollable(history: &mut InputHistory) -> String{
                 if gstate == 2 {
                     gstate = 0; 
                     his_index -= 1;
-                    let val = history.get_index(his_index);
-                    if let Some(valv) = val{
-                        scroll_action(&mut res, valv, &mut pos);
-                    }
+                    scroll_action(&mut res, &mut pos, &history, his_index);
                 }
                 else { typed_char(66, &mut res, &mut gstate, &mut hoen_state, &mut pos); }
             }
